requirements.txt
aioftp==0.21.4
<<<<<<< HEAD
bigtree==0.12.5
certifi==2024.8.30
cffi==1.15.1
cramjam==2.9.0
dateparser==1.2.0
dbfread==2.0.7
elasticsearch==7.16.2
fastparquet==0.8.3
fsspec==2024.10.0
greenlet==3.1.1
humanize==4.11.0
loguru==0.6.0
numpy==1.26.2
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pycparser==2.21
pyreaddbc==1.2.0
pysus==0.11.0
=======
alembic==1.14.0
asttokens==2.4.1
bigtree==0.12.5
certifi==2024.8.30
cffi==1.15.1
comm==0.2.2
compose-go==2.30.3
cramjam==2.9.0
dateparser==1.2.0
dbfread==2.0.7
debugpy==1.8.9
decorator==5.1.1
elasticsearch==7.16.2
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.11.0
fsspec==2024.10.0
greenlet==3.1.1
humanize==4.11.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.6.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.21
Pygments==2.18.0
pyreaddbc==1.2.0
pysus==0.14.3
>>>>>>> 495b7d3 (upload SIH e SIA em andamento)
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
<<<<<<< HEAD
regex==2024.11.6
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.64.0
=======
pyzmq==26.2.0
regex==2024.11.6
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tornado==6.4.2
tqdm==4.64.0
traitlets==5.14.3
>>>>>>> 495b7d3 (upload SIH e SIA em andamento)
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
urllib3==1.26.20
urwid==2.6.16
wcwidth==0.2.13
wget==3.2
